requirements.txt
beautifulsoup4
selenium
pillow
lxml
colorama